General Update
The Fourth Version of my simple-but-not-so-humble calculator! lots of foreign concepts in this one!

for HTML, I completely re-structured the document flow, making every section separate. the title, the buttons for the modes, the report with the input, output and the history. (this will be the new structure for the site, for now)

for CSS, I added LOTS and LOTS of grids, an insane amount, all done using ordered lists, I changed the color scheme to something darker, I learned how to use the basics of margin and padding and how to inspect them using devtools which really helped!

I still use absolute position and I dropped the old design.

this new version is still not functional sadly... but, I will add the JS and follow on the output in the next commit!

for JS, I did nothing :D